Move ValidationResult to sip.go (and its tests to sip_test.go); update Combine implementation.
