+++ livekit/sip_validation.go
@@ -19,8 +19,6 @@ import (
 	fmt "fmt"
 	"strconv"
 	"strings"
-
-	"github.com/livekit/protocol/logger"
 )
 
 // RFC 3261 compliant validation functions for SIP headers and messages
@@ -216,67 +214,6 @@ var nameAddrHeaders = map[string]bool{
 
 // Deprecated: has no effect. Use ValidationResult method variants instead.
 func SetSoftFailureReporter(reporter func(err error)) {}
-
-// ValidationResult holds information about something that's been validated.
-type ValidationResult struct {
-	err      error
-	softErrs []error
-}
-
-// OK returns whether or not an error was seen.
-func (vr ValidationResult) OK() bool {
-	return vr.err == nil
-}
-
-// Error returns the error associated with the result.
-func (vr ValidationResult) Error() error {
-	return vr.err
-}
-
-// SoftErrors returns errors that can probably be ignored.
-func (vr ValidationResult) SoftErrors() []error {
-	return vr.softErrs
-}
-
-// Combine returns a new validation result that is comprised of this result and
-// the other result.
-func (vr ValidationResult) Combine(other ValidationResult) ValidationResult {
-	var err error
-	if vr.err != nil {
-		err = vr.err
-	} else {
-		err = other.err
-	}
-	softErrs := append(vr.softErrs, other.softErrs...)
-	return ValidationResult{err, softErrs}
-}
-
-// WithError returns a new ValidationResult with the given error.
-func (vr ValidationResult) WithError(err error) ValidationResult {
-	return ValidationResult{err, vr.softErrs}
-}
-
-// LogSoftErrors logs warnings for each of the result's soft errors and
-// returns the result's associated error, if any.
-func (vr ValidationResult) LogSoftErrors(l logger.Logger) error {
-	for _, err := range vr.softErrs {
-		l.Warnw("soft validation failure: %w", err)
-	}
-	return vr.err
-}
-
-// LogUnlikelySoftErrors is a variant of LogSoftErrors.
-func (vr ValidationResult) LogUnlikelySoftErrors(l logger.UnlikelyLogger) error {
-	for _, err := range vr.softErrs {
-		l.Warnw("soft validation failure: %w", err)
-	}
-	return vr.err
-}
-
-// ValidationFailure returns a new ValidationResult.
-func ValidationFailure(err error) ValidationResult {
-	return ValidationResult{err, nil}
-}
 
 // ValidateHeaderName validates a SIP header name per RFC 3261 Section 25.1
 func ValidateHeaderName(name string, restrictNames bool) error {
